Replaced use of arrow functions with static
This causes an issue in Firefox 68, which is the current mobile build of FF on android.
It might have also caused issues on other browsers, so this unnecessary small thing can
be changed to improve support across browsers.

It took a lot of effort to debug this issue since I found it on mobile yesterday.
- Tried remote debugging, but version mismatches had issues with v68
- Tried FF Preview on mobile, but it didn't have the issue
- Installed v68 on desktop, but remote debugging didn't work
- Used v68 on desktop to reproduce the issue, and it was a simple fix.